+++ src/main/java/Projetos_Modulo_inicial/TaskManager/WorkTask.java
@@ -4,7 +4,7 @@ public class WorkTask extends Task {
 
     private String worTask;
 
-    public WorkTask(int id, String descricao, String worTask) {
+    public WorkTask(long id, String descricao, String worTask) {
         super(id, descricao);
         this.worTask = worTask;
     }
